tb9660: fix handling of >4GB files on platforms where long is 64-bit
Back in 2014 I only tested on CD ISOs, which max out at 700MB. ISOs can,
of course, get much larger. It's hard to fix this (while staying within
the capabilities of stdio.h) generically without introducing a build system
which can detect fseeko/off_t (POSIX) or _fseeki64 (MSVC). However, we
can at least make it work for the straightforward case where long is big
enough.

Fixes #1

+++ tb9660.c
@@ -2,7 +2,7 @@
  *          systems
  *
  * SPDX-License-Identifier: LicenseRef-ISC0
- * SPDX-FileCopyrightText: Copyright © 2014 Erin Shepherd
+ * SPDX-FileCopyrightText: Copyright © 2014, 2025 Erin Shepherd
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted without restriction.
@@ -28,7 +28,7 @@ l9660_file file;
 
 static bool readsect(l9660_fs *fs, void *buf, uint32_t sector)
 {
-    if (fseek(isof, 2048 * sector, SEEK_SET)) {
+    if (fseek(isof, 2048 * (long)sector, SEEK_SET)) {
         fprintf(stderr, "Reading sector %u\n", sector);
         perror("fseek");
         return false;
